Rename naive_bayse module and related files
